refactor: unvalidated email send to own file
